+++ CHANGELOG.md
@@ -7,7 +7,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ### Added
 - Added CPPNImage class implementing Compositional Pattern Producing Networks
 - Added Imaging functionality to Image objects, can now call to_pyplot / to_file directly on an image
+- Added the PyramidAscent class which can be used to progressively increase image scale as in deep dream
 ### Changed
+- Renamed BasicAscent to Ascent
 ### Deprecated
 ### Removed
 ### Fixed